+++ source/DSCResources/MSFT_xCluster/MSFT_xCluster.psm1
@@ -43,7 +43,7 @@ function Get-TargetResource
         [System.String[]]
         $IgnoreNetwork,
 
-        [Parameter(Mandatory = $true)]
+        [Parameter()]
         [System.Management.Automation.PSCredential]
         $DomainAdministratorCredential
     )
@@ -56,10 +56,13 @@ function Get-TargetResource
         $errorMessage = $script:localizedData.TargetNodeDomainMissing
         New-InvalidOperationException -Message $errorMessage
     }
-
+    $context = $null
     try
     {
-        ($oldToken, $context, $newToken) = Set-ImpersonateAs -Credential $DomainAdministratorCredential
+        if ($PSBoundParameters.ContainsKey('DomainAdministratorCredential'))
+        {
+            ($oldToken, $context, $newToken) = Set-ImpersonateAs -Credential $DomainAdministratorCredential
+        }
 
         $cluster = Get-Cluster -Name $Name -Domain $computerInformation.Domain
         if ($null -eq $cluster)
@@ -87,6 +90,7 @@ function Get-TargetResource
         IgnoreNetwork                 = $IgnoreNetwork
         DomainAdministratorCredential = $DomainAdministratorCredential
     }
+
 }
 
 <#
@@ -136,7 +140,7 @@ function Set-TargetResource
         [System.String[]]
         $IgnoreNetwork,
 
-        [Parameter(Mandatory = $true)]
+        [Parameter()]
         [System.Management.Automation.PSCredential]
         $DomainAdministratorCredential
     )
@@ -168,7 +172,11 @@ function Set-TargetResource
 
     try
     {
-        ($oldToken, $context, $newToken) = Set-ImpersonateAs -Credential $DomainAdministratorCredential
+        $context = $null
+        if ($PSBoundParameters.ContainsKey('DomainAdministratorCredential'))
+        {
+            ($oldToken, $context, $newToken) = Set-ImpersonateAs -Credential $DomainAdministratorCredential
+        }
 
         if ($bCreate)
         {
@@ -302,7 +310,7 @@ function Test-TargetResource
         [System.String[]]
         $IgnoreNetwork,
 
-        [Parameter(Mandatory = $true)]
+        [Parameter()]
         [System.Management.Automation.PSCredential]
         $DomainAdministratorCredential
     )
@@ -320,7 +328,11 @@ function Test-TargetResource
 
     try
     {
-        ($oldToken, $context, $newToken) = Set-ImpersonateAs -Credential $DomainAdministratorCredential
+        $context = $null
+        if ($PSBoundParameters.ContainsKey('DomainAdministratorCredential'))
+        {
+            ($oldToken, $context, $newToken) = Set-ImpersonateAs -Credential $DomainAdministratorCredential
+        }
 
         $cluster = Get-Cluster -Name $Name -Domain $ComputerInfo.Domain